Kernel video engine altered, support for 136cy sync from IT dropped.
This is for the normal video lines, where it might be useful to run the
sync from interrupt for a small mixer routine (vsync mixer). It is dropped
since the controller readings would interfere with the interrupt based
rising of the sync edge. For FoaD this feature isn't necessary anyway.