Further progress towards a complete game.
Code-wise the game should be mostly complete except for some bugs from
untested actors. I will determine later whether I will add a catapult.

The first two levels are complete, the environment of the second two is
added along with fixing related typos in the assets.

The layout and sizes in the resources should be finalized now (hopefully
if I decide on adding a catapult, that won't upset it much in the sprite
layouts). Also lots of size trimmings were performed in actors.

End game cutscene is also added.