Reimplemented the resources file in assembler.
No functional changes. Binary result should be the same if compiled by the
same version of GCC (unaffected by the resource ordering problem). Note that
a Linker Script had to be added as for some obscure reason the linker
parameter when the assembly resource is used, produces the section with no
Allocate flag, thus missing from the output.

Tested by a complete playthrough, game compiles and runs fine with avr-gcc
5.4.0